Deploy to AWS Lambda using serverless toolkit
Since we're using AWS for our general deployment, and to avoid having the deployment of this bot to a specific Heroku account, I've rewritten to target AWS Lambda as the deployment target.

Replace the express.js server with a handler.js script.

Use `rollup` to merge all scripts into a single deployable file.

Had to change the ESLint and stylelint config files to JavaScript files so that they can be imported and rolled up with the rest of the code, since a serverless setup can't read from a filesystem.