Update response codes for unprocessed requests
Avoid responding with a 4xx status on unsupported actions, since it will turn up as errors in the GitHub request logs. The request was a correctly formated GitHub pull-request webhook request, but we just don't want to perform linting on the particular action (like 'closed').

Also, use 412 Precondition Failed when the correct headers are not present.